+++ packages/api/src/modules/task/dto/UpdateTaskInput.ts
@@ -1,5 +1,5 @@
 import { TaskPriority, TaskStatus } from "@dewo/api/models/Task";
-import { Field, InputType, Int } from "@nestjs/graphql";
+import { Field, Float, InputType } from "@nestjs/graphql";
 import GraphQLUUID from "graphql-type-uuid";
 import { UpdateTaskRewardInput } from "./UpdateTaskRewardInput";
 import { UpdateTaskReviewInput } from "./UpdateTaskReviewInput";
@@ -37,7 +37,7 @@ export class UpdateTaskInput {
   @Field(() => [GraphQLUUID], { nullable: true })
   public ownerIds?: string[];
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   public storyPoints?: number;
 
   @Field(() => TaskStatus, { nullable: true })
